requirements.txt
opencv-python
natsort
pillow
matplotlib
